sched scan: Avoid scan if no enabled networks
This fixes a bug where the supplicant starts with no networks in the conf
and a scan is initiated through wpa_cli.
After receiving the results and finding no match sched scan
was requested but when initiated with no networks in filters and no
networks for active scan it triggers a periodic scan result event
with no BSSs every scan which causes the scan results in the supplicant
to be deleted.

Signed-off-by: Eyal Shapira <eyal@wizery.com>