sched scan: Avoid scan if no enabled networks
There's no point to sched scan when there's no enabled networks.
This also fixes a bug where the supplicant would reset
its scan results in case such a sched scan was initiated
and therefore would cause a "flickering" network list effect on Android.
(For a moment you see the network list and the next it's empty)

Signed-off-by: Eyal Shapira <eyal@wizery.com>