nl80211: Implement some private cmds
Implement the LINKSPEED and RSSI/RSSI-APPROX private commands.

Upstream implementation sends the HANGED state event on
private command failure. Unimplemented commands were
sent to the driver as-is, causing the HANGED event to be sent
and cause a recovery. Private command handling was modified so
that only implemented commands are sent to the driver and
the sending of the HANGED event was removed.

[Arik - The "linkspeed" driver command incorrectly reported the
rate in kbps. Fix this]

Signed-off-by: Arik Nemtsov <arik@wizery.com>
Signed-off-by: Yotam Rubin <yotam@wizery.com>