P2P: ignore dialog token of p2p negotiation confirmation
The P2P spec is pretty crappy, and doesn't take care
of the case in which the dialog token of the p2p
negotiation confirmation is wrong - in such case the
sending side will start a group, while the remote side
will reject it (due to incorrect dialog token).

Workaround it by ignoring the dialog token of the
p2p negotiation confirmation packets.

Signed-off-by: Eliad Peller <eliad@wizery.com>