p2p: consider age for the p2p scan results
Cfg80211 caches the scan results according the
channel number. Due to the 15 sec aging this
might cause the user mode to see more than one
scan result with the same BSSID, e.g. - one
scan result for the p2p-device and one for the
p2p-GO (once it's enabled).

Fix this by updating the device entry only if
the new peer entry is newer than the one
previously stored.

Signed-off-by: Yoni Divinsky <yoni.divinsky@ti.com>
Signed-off-by: Victor Goldenshtein <victorg@ti.com>
Signed-off-by: Igal Chernobelsky <igalc@ti.com>