P2P: Start blocked p2p find even if scan was on another interface (BRCM)
p2p find can now defer starting the p2p scan due to an ongoing
scan on a different interface. Relying on p2p_cb_on_scan_complete
which is a per interface flag is wrong here and we won't resume
the p2p find in case the scan wasn't on the same interface (which is
the case in Android JB).
Fix this by relying on the global p2p state instead.

(This description is based on analysis of the code as the original
BRCM commit doesn't contain any useful information)

Signed-off-by: Eyal Shapira <eyal@wizery.com>