Fix use-after-free when removing all networks (UPSTREAM)
commit 59ff6653
"Fix REMOVE_NETWORK to not run operations with invalid current_ssid"
fixed the removal of a single network but the same use-after-free
issue exists when removing all networks.

As part of the fix refactor ctrl_iface_remove_network to use
wpa_supplicant_remove_network which actually does the work.
This should help with fixing the use-after-free issue also when
remove network is done via dbus.

Signed-off-by: Eyal Shapira <eyal@wizery.com>