Increase channel delay to 100ms
There is an inherent flaw in the p2p protocol design where an ACK loss right
before channel switch leads to peers being out of sync.

A work around of 50ms was added persistent case, but it turns out in heavy lossy
conditions that is not enough. Increase it to 100ms.

Experimental evaluation showed 100ms improves the reliability of p2p persistence
reinvocation.

Bug: 7359500
Change-Id: I73ab1b64e32b87101e95b0e487c90818c2c0aaf3